% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_se_t.R
\name{Delta_se_t}
\alias{Delta_se_t}
\title{A function to compute change points based on the  MNP method}
\usage{
Delta_se_t(y, z, s, e, t, h)
}
\arguments{
\item{y}{a matrix containing the data. Row correspond to different time points and columns to different variables}

\item{z}{a copy of the matrix y, it can be y itself}

\item{e}{}

\item{h}{bandwith parameter}
}
\description{
A function to compute change points based on the  MNP method
}
